# modules/02-backend-core/contract/mvp.openapi.yaml
openapi: 3.1.0
info:
  title: 02-backend-core MVP API (Draft)
  version: 0.1.0-draft
  description: |
    草案：用于 `modules/99-hub` 定稿的输入（非最终版本）。
servers:
  - url: http://localhost:3000

tags:
  - name: auth
  - name: questionBank
  - name: questions
  - name: videos
  - name: svip
  - name: activation
  - name: me
  - name: records
  - name: content
  - name: billing
  - name: progress
  - name: wrongs
  - name: favorites

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
          description: 统一错误码（需 99-hub 定稿枚举）
          examples: [UNAUTHORIZED, FORBIDDEN, INVALID_ARGUMENT, NOT_FOUND, INTERNAL]
        message:
          type: string
        requestId:
          type: string
          nullable: true

    AuthWechatRequest:
      type: object
      additionalProperties: false
      required: [code]
      properties:
        code:
          type: string
          description: 微信登录 code

    AuthWechatResponse:
      type: object
      additionalProperties: false
      required: [token, expiresIn]
      properties:
        token:
          type: string
          description: 访问 token（草案：Authorization Bearer 传递）
        expiresIn:
          type: integer
          minimum: 1
          description: token 有效期（秒）
        userId:
          type: string
          nullable: true
        isNewUser:
          type: boolean
          nullable: true

    Subject:
      type: object
      additionalProperties: false
      required: [id, name]
      properties:
        id: { type: string }
        name: { type: string }

    Package:
      type: object
      additionalProperties: false
      required: [id, subjectId, name]
      properties:
        id: { type: string }
        subjectId: { type: string }
        name: { type: string }

    Chapter:
      type: object
      additionalProperties: false
      required: [id, packageId, name]
      properties:
        id: { type: string }
        packageId: { type: string }
        name: { type: string }

    Question:
      type: object
      additionalProperties: false
      required: [id, chapterId, stem]
      properties:
        id: { type: string }
        chapterId: { type: string }
        stem: { type: string }
        analysis:
          type: string
          nullable: true

    AnswerSubmitRequest:
      type: object
      additionalProperties: false
      required: [chapterId, answers]
      properties:
        chapterId: { type: string }
        answers:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [questionId, value]
            properties:
              questionId: { type: string }
              value: { type: string }

    AnswerSubmitResponse:
      type: object
      additionalProperties: false
      required: [submitted]
      properties:
        submitted: { type: integer, minimum: 0 }

    ProgressResponse:
      type: object
      additionalProperties: false
      required: [packageId, answeredCount, correctCount, totalCount, updatedAt]
      properties:
        packageId: { type: string }
        answeredCount: { type: integer, minimum: 0 }
        correctCount: { type: integer, minimum: 0 }
        totalCount: { type: integer, minimum: 0 }
        updatedAt:
          type: string
          format: date-time

    WrongItem:
      type: object
      additionalProperties: false
      required: [questionId, chapterId, wrongAt]
      properties:
        questionId: { type: string }
        chapterId: { type: string }
        wrongAt:
          type: string
          format: date-time

    FavoriteUpsertRequest:
      type: object
      additionalProperties: false
      required: [questionId]
      properties:
        questionId: { type: string }

    FavoriteUpsertResponse:
      type: object
      additionalProperties: false
      required: [favorited, questionId]
      properties:
        favorited: { type: boolean }
        questionId: { type: string }

    FavoriteDeleteResponse:
      type: object
      additionalProperties: false
      required: [unfavorited, questionId]
      properties:
        unfavorited: { type: boolean }
        questionId: { type: string }

    Video:
      type: object
      additionalProperties: false
      required: [id, title]
      properties:
        id: { type: string }
        title: { type: string }
        url:
          type: string
          nullable: true

    VideoProgressUpsertRequest:
      type: object
      additionalProperties: false
      required: [videoId, progress]
      properties:
        videoId: { type: string }
        progress:
          type: number
          minimum: 0
          maximum: 1
          description: 播放进度（0~1）

    VideoProgressResponse:
      type: object
      additionalProperties: false
      required: [videoId, progress]
      properties:
        videoId: { type: string }
        progress:
          type: number
          minimum: 0
          maximum: 1

    SvipBindRequest:
      type: object
      additionalProperties: false
      required: [code]
      properties:
        code:
          type: string
          description: SVIP/激活码（草案字段名，需 99-hub 定稿）

    SvipBindResponse:
      type: object
      additionalProperties: false
      required: [bound]
      properties:
        bound: { type: boolean }

    ActivationRedeemRequest:
      type: object
      additionalProperties: false
      required: [code]
      properties:
        code: { type: string }

    ActivationRedeemResponse:
      type: object
      additionalProperties: false
      required: [redeemed]
      properties:
        redeemed: { type: boolean }

    MeEntitlementsResponse:
      type: object
      additionalProperties: false
      required: [active, subjects, packages]
      properties:
        active: { type: boolean }
        subjects:
          type: array
          items: { type: string }
        packages:
          type: array
          items: { type: string }

    LearningRecord:
      type: object
      additionalProperties: false
      required: [id, chapterId, updatedAt]
      properties:
        id: { type: string }
        chapterId: { type: string }
        updatedAt:
          type: string
          format: date-time

    Banner:
      type: object
      additionalProperties: false
      required: [id, title]
      properties:
        id: { type: string }
        title: { type: string }
        imageUrl:
          type: string
          nullable: true
        link:
          type: string
          nullable: true

    NewsItem:
      type: object
      additionalProperties: false
      required: [id, title, publishedAt]
      properties:
        id: { type: string }
        title: { type: string }
        summary:
          type: string
          nullable: true
        publishedAt:
          type: string
          format: date-time

    NewsDetail:
      type: object
      additionalProperties: false
      required: [id, title, content, publishedAt]
      properties:
        id: { type: string }
        title: { type: string }
        content: { type: string }
        publishedAt:
          type: string
          format: date-time

    Recommendation:
      type: object
      additionalProperties: false
      required: [id, title, type, refId]
      properties:
        id: { type: string }
        title: { type: string }
        type: { type: string }
        refId: { type: string }

    BillingPlan:
      type: object
      additionalProperties: false
      required: [id, name, priceCents, currency]
      properties:
        id: { type: string }
        name: { type: string }
        priceCents: { type: integer, minimum: 0 }
        currency: { type: string }

    StudyRecordUpsertRequest:
      type: object
      additionalProperties: false
      required: [chapterId, payload]
      properties:
        chapterId: { type: string }
        payload:
          type: object
          description: 草案：学习记录内容（后续可拆字段）

    StudyRecordResponse:
      type: object
      additionalProperties: false
      required: [id]
      properties:
        id: { type: string }

  responses:
    BadRequest:
      description: 参数错误
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: 未认证
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: 无权限
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

paths:
  /auth/wechat:
    post:
      tags: [auth]
      summary: 微信登录（code → token）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthWechatRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthWechatResponse"
        "400":
          $ref: "#/components/responses/BadRequest"

  /subjects:
    get:
      tags: [questionBank]
      summary: 获取科目列表
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subject"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /packages:
    get:
      tags: [questionBank]
      summary: 获取题库包列表
      security:
        - bearerAuth: []
      parameters:
        - name: subjectId
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Package"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /chapters:
    get:
      tags: [questionBank]
      summary: 获取章节列表
      security:
        - bearerAuth: []
      parameters:
        - name: packageId
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Chapter"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /questions:
    get:
      tags: [questions]
      summary: 获取题目列表
      security:
        - bearerAuth: []
      parameters:
        - name: chapterId
          in: query
          required: true
          schema: { type: string }
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 200 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Question"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /answers/submit:
    post:
      tags: [records]
      summary: 提交作答（草案）
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnswerSubmitRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnswerSubmitResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /progress:
    get:
      tags: [progress]
      summary: 获取做题进度（草案）
      security:
        - bearerAuth: []
      parameters:
        - name: packageId
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgressResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /wrongs:
    get:
      tags: [wrongs]
      summary: 获取错题列表（草案）
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WrongItem"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /favorites:
    post:
      tags: [favorites]
      summary: 收藏题目（草案）
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FavoriteUpsertRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FavoriteUpsertResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
    delete:
      tags: [favorites]
      summary: 取消收藏（草案）
      security:
        - bearerAuth: []
      parameters:
        - name: questionId
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FavoriteDeleteResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /videos:
    get:
      tags: [videos]
      summary: 获取视频列表
      security:
        - bearerAuth: []
      parameters:
        - name: packageId
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Video"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /videos/progress:
    get:
      tags: [videos]
      summary: 获取视频进度
      security:
        - bearerAuth: []
      parameters:
        - name: videoId
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoProgressResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [videos]
      summary: 更新视频进度
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VideoProgressUpsertRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoProgressResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /svip/bind:
    post:
      tags: [svip]
      summary: SVIP 绑定（草案）
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SvipBindRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SvipBindResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

  /activation/redeem:
    post:
      tags: [activation]
      summary: 激活码兑换（草案）
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivationRedeemRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivationRedeemResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /me/entitlements:
    get:
      tags: [me]
      summary: 获取我的权益（草案）
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeEntitlementsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /me/learning-records:
    get:
      tags: [me]
      summary: 获取学习记录（草案）
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LearningRecord"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /content/banners:
    get:
      tags: [content]
      summary: 获取 banner（草案）
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Banner"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /content/news:
    get:
      tags: [content]
      summary: 获取资讯列表（草案）
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NewsItem"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /content/news/{id}:
    get:
      tags: [content]
      summary: 获取资讯详情（草案）
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /content/recommendations:
    get:
      tags: [content]
      summary: 获取推荐内容（草案）
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recommendation"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /billing/plans:
    get:
      tags: [billing]
      summary: 获取套餐列表（草案）
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BillingPlan"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /records:
    get:
      tags: [records]
      summary: 获取学习记录（草案）
      security:
        - bearerAuth: []
      parameters:
        - name: packageId
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudyRecordResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [records]
      summary: 写入学习记录（草案）
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudyRecordUpsertRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyRecordResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"